requirements.txt
click
msgpack-python
pyzmq
tornado
